+++ tests/correctness/targets/CustomCommandSecretHiding/testcmd.py
@@ -1,0 +1,2 @@
+import sys
+print('Command completed successfully - %s args'%len(sys.argv[1:]))